requirements.txt
pinecone-client==2.2.1
langchain==0.0.123
openai
chromadb
pandas
numpy
